+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.3.5] - 2018-11-07
+### Added
+- **`MessagesContext`**
+  - Create context to pass iframe locale messages to `PageEditor`.
+- **`EditorProvider`**
+  - New component to wrap `EditorProvider` with `MessagesContext.Consumer`.
+- **`PageEditor`**
+  - static method `getCustomMessages`
+  - `Message.Provider`
+
 ## [2.3.4] - 2018-11-06
 ### Changed
 - **`pages.json`**